search improved

+++ mysearchproapp/models.py
@@ -7,14 +7,12 @@ class Category(models.Model):
     def __str__(self):
         return self.name
 
-class Product(models.Model): 
-    id = models.BigAutoField(primary_key=True) 
-    name = models.CharField(max_length=255) 
-    description = models.TextField() 
-    price = models.DecimalField(max_digits=10, decimal_places=2) 
+class Product(models.Model):
+    id = models.BigAutoField(primary_key=True)
+    name = models.CharField(max_length=255)
+    description = models.TextField()
+    price = models.IntegerField()
     category = models.ForeignKey(Category, on_delete=models.CASCADE, default=1)
 
 def __str__(self):
     return self.name
-
-
